tvafe: change vbi addr variable to long [1/1]
PD#SWPL-136297

Problem:
vbi addr variable is int cause abnormal

Solution:
change vbi addr variable to long

Verify:
t5m

+++ drivers/amlogic/media/vin/tvin/tvafe/tvafe_cvd.c
@@ -192,7 +192,7 @@ static unsigned int noise3;
 /* test */
 #define NOSTD_DEBUG_PRINT_CNT     100
 
-unsigned int vbi_mem_start;
+unsigned long vbi_mem_start;
 
 static int acd_2d_adjust = 0x94;
 module_param(acd_2d_adjust, int, 0644);
